backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
sqlmodel
psycopg2-binary
Pillow